update assets for get iamge and delete iamge from $home, not use project/images/

+++ src/main/java/com/project/repository/AnimeRepository.java
@@ -2,7 +2,7 @@ package com.project.repository;
 
 import com.project.entity.Anime;
 
-public class AnimeRepository extends JpaRepository<Anime, Integer> {
+public class AnimeRepository extends JpaRepository<Anime, Long> {
     public AnimeRepository() {
         super();
     }
